refactor: 테스트를 위한 변수들 제거 전 리팩토링
- domain package, ui package 추가 및 이동
- Number 객체 toString() 메서드 추가
- 메서드 기능내역에 맞게 변경 (isNullOrZero -> isNullOrEmpty)
